requirements.txt
requests==2.32.3
python-dotenv==1.0.1
python-magic-bin==0.4.14
pandas==2.2.3